Nearby Zipcode - Marketing
During marketing analysis of campaign, we often come across the problem to find zipcodes. Zipcodes used by marketing agency to position billboards or advertisements. We usually require a cluster of zipcodes which have a near proximity to the zipcodes where the advertisement is/was positioned. 
In this code we use a 10 mile proximity of zipcodes from the origination target zipcodes to perform our analysis.